this is the first step in merging into my fork of the Microsoft-controlled DotNetty repo the work done by wizard872 in his fork of the DotNetty code.  i would use his/her code directly, apart from the fact that it appears that many packages were removed and the last commit was in June 0f 2020.  as indicated in the file called 220506_dotNetty_analysis.txt a diff of the directories indicates that the changes are rather narrow in scope and number, and so my hope is that merging in these changes won't prove problematic.
once this is done, CORE:EPIC will have a DotNetty version for the client-side that is
compatible with Unity.  my hope is that this will mean that code and tests written for
the Java side of the world (the server) can be easily migrated to this compatible library.

keeping this in a branch for now, but we will merge it with the main once we know it works.

to be clear, i've made no changes to the code as yet.  just copied the changed files
into separate subdirectories where i can start working on them in Rider.  i've also performed
a build of the Microsoft code, which will not build within Rider's IDE, but will build
on the command-line.

that's it for now...

...END-OF-LINE...