ok, so the changes from wizard872 have been pulled in properly.
apart from the significant changes to:

DotNetty.Transport/Channels/Sockets/SocketDatagramChannel.cs

it seems most of the changes are to replace .Net's Thread class
with something homegrown.  if i believe the statements in this
forum thread:

https://forum.unity.com/threads/anyone-successfully-use-dotnetty-in-this-build.456984/

this is because Unity isn't multi-threaded.  the homegrown class is:

DotNetty.Common/Concurrency/XThread.cs

looks like it's missing some parts of the Thread class that were
expected by other parts of the original DotNetty code.  i've worked
around this for now, but i'll need to get back in there to put
something together that's a bit less kludgey.

something else from that forum thread is an argument that DotNetty
isn't really necessary because of some of the async I/O methods in
C#.  pretty sure there's not much truth to that considering what
i have already had to go through to get I/O to work with the
CORE gateway.

in any case, the wizard872 code has been applied into the appropriate
places, and the final result compiles successfully on both the command-line
and in the Rider IDE (not sure how that happened)...

next step is to build a primitive Unity project with some echo code and
get back to work on the server.

...END-OF-LINE...