amazing what a couple of small changes will do (and won't).
the goal for this commit was just to alter the code enough to get a set of DLLs
imported into Unity that didn't crash on import.  that has been accomplished, but
with a compromise: we might have to give up using DotNetty.Protobuf in this first
design.

the story behind the reasons is long, but the upshot is that the .Net build process
does not allow for the manipulation of dependencies of dependencies.  this is something
i take for granted coming out of Java, where through Maven you can suppress an old
version of a package in a dependency, so that only a newer one has precedence.

in .Net, the two competing packages just collide.  the solution according to Microsoft:

“To resolve incompatibilities, do one of the following:

* Retarget your project to a framework that is supported by the packages you want to use.

* Contact the author of the packages and work with them to add support for your chosen
framework. Each package listing page on nuget.org has a Contact Owners link for this purpose.”

so, if the author is Google and the package is Google.Protobuf, good luck getting them to
work with you...and i don't say this out of hand.  i went and read some of the comments.

so, as i see it i have two possible solutions:

1) clone the Google.Protobuf repo, make the changes, build it myself, and then figure out
how to pull it into DotNetty.Protobuf,

or

2) just do without DotNetty.Protobuf.

tonight, i'm choosing the latter, but who knows how i'll feel tomorrow.

on the Unity side, i'll proceed with all the DotNetty DLLs except DotNetty.Protobuf.
this is a more comprehensive set of DLLs than what wizard872 used, so hopefully they all work.

one thing to note is that there are a couple of projects that have done PRs and such to the
Google code with suggestions on how Google might improve the code for compatibility with
Unity.  will have to take a look at those.

now for the stuff that did work...

* Unity had multiple complaints about the import, but after adjusting the API Compatibility Level
to .Net 4.x, and the Scripting Backend to IL2CPP, most went away.  those that did not involved
upgrading dependencies.

* Google.Protobuf was upgraded to version 3.20.1 (which doesn't solve the ultimate problem).

* System.Diagnostics.DiagnosticSource was added to the DotNetty.Common package at version 5.0.0.

because we're on the newer version of NuGet, these changes were made in the .csproj files below.

	modified:   src/DotNetty.Codecs.Protobuf/DotNetty.Codecs.Protobuf.csproj
	modified:   src/DotNetty.Common/DotNetty.Common.csproj

* the last change was just adding a directory to coalesce the DLLs before bringing them into Unity.
i don't want to commit that to the repo, so .gitignore got a visit.

	modified:   .gitignore

that's it...END-OF-LINE...